pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "decorator>=5.1.1",
    "imageio>=2.37.0",
    "instagrapi>=2.1.3",
    "moviepy>=2.1.2",
    "numpy>=2.2.2",
    "openai>=1.61.1",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium==4.16.0",
    "slack-sdk>=3.34.0",
    "trafilatura>=2.0.0",
    "undetected-chromedriver==3.5.3",
    "webdriver-manager==4.0.1",
]
